+++ src/consts.py
@@ -1,0 +1,2 @@
+PRAGALERIA_UPLOAD_URL = 'http://pragaleria.pl/wp-content/uploads/'
+PRAGALERIA_AUCTIONS_URL = 'http://pragaleria.pl/aukcje-wystawy/'
